Add code blocks to scratch pad
Code blocks are the building blocks of a scratch pad. Later they will
shown in a CodeMirror editor, but for the moment the basic
functionallity can be implemented with a plain text area. This solution
is pretty good, since CodeMirror behaves as a something hooked on top of
a textarea object.