Load dummy data from API
The API now serves the dummy data via HTTP and the UI is fetching it
from there. During development a proxy mounts the API into the UI.

Sadly, the isLoading property in the state causes an infinite loop
loading the data again and again. For the moment, this functionality is
disabled.